Refactored the mock-server
