+++ package.json
@@ -34,6 +34,7 @@
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.2",
     "eslint-plugin-react": "^6.3.0",
+    "express-boom": "^1.0.0",
     "express-jwt": "^5.1.0",
     "faucet": "0.0.1",
     "gulp": "^3.9.1",
